// app/generated/api/models/attributeType.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Pricing Backend API
 * REST API for managing branches, regions, products, and pricing plans.
 * OpenAPI spec version: 1.0.0
 */

export type AttributeType = typeof AttributeType[keyof typeof AttributeType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AttributeType = {
  TEXT: 'TEXT',
  DECIMAL: 'DECIMAL',
  INTEGER: 'INTEGER',
  DATE: 'DATE',
  BOOLEAN: 'BOOLEAN',
} as const;